+++ randint/src/utils.h
@@ -1,6 +1,8 @@
 #ifndef		__UTILS_H
 #define		__UTILS_H
 
-void int_to_str(long, char *, int);
+#include <stdint.h>
+
+void int_to_str(uint32_t, char *, int);
 
 #endif
